% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hPAA.R
\name{hPAA}
\alias{hPAA}
\title{Hierarchical Principal Amalgamation Analysis with/without Taxonomic Tree Guidance}
\usage{
hPAA(X, taxonomy = NULL, 
     method = c("Simpson", "Shannon", "Bray-Curtis", "Weighted UniFrac"), 
     W = NULL, strong = FALSE)
}
\arguments{
\item{X}{a data matrix of compositions (\eqn{n \times p}).}

\item{taxonomy}{a vector of taxonomic tree structure related to the compositions
which is used to conduct tree-guided hierarchical analysis.
Each element of the vector denotes the taxonomic ranks of the components, 
separated by semicolon, in the format of mothur taxonomy output (\eqn{p \times 1}).}

\item{method}{the diversity index used to conduct the hierarchical amalgamation.}

\item{W}{a vector of weight for each sample in measure of diversity (\eqn{n \times 1}).}

\item{strong}{indicator of tree-guided hierarchical approach, default is weak hierarchy taxonomy
method. The parameter is valid only if taxonomic tree matrix is provided.}
}
\value{
An object of class "hPAA" which describes the tree produced by the amalgamation process
\item{merge}{a \eqn{p-1} by 2 matrix. Row \eqn{i} of merge describes the merging of components at step \eqn{i} 
             of the amalgamation. If an element \eqn{j} in the row is negative, then observation \eqn{-j} was merged 
             at this stage. If \eqn{j} is positive then the merge was with the new composition formed at the (earlier) stage 
             \eqn{j} of the algorithm. Thus negative entries in merge indicate agglomerations of singletons, and positive 
             entries indicate agglomerations of non-singletons.}
\item{imerge}{a \eqn{p-1} by 2 matrix. Row \eqn{i} of merge describes the merging of compositions at step \eqn{i} 
              of the amalgamation, as the indexes of columns in the original compositional data matrix. }
\item{height}{a set of \eqn{p-1} real values. The value of the diversity measure at each amalgamation step.}
\item{crit}{a set of \eqn{p-1} real values. The decrease of diversity meaure after amalgamation at each step.}
\item{order}{a vector giving the permutation of the original compositions suitable for plotting, 
             in the sense that a hierarchical dendrogram using this ordering and matrix merge will not have 
             crossings of the branches.}
\item{labels}{labels for each of the components being amalgamated.}
\item{diversity.method}{the measure of diversity used for conducting the analysis.}
\item{hierarchical.method}{the approach used for conducting the analysis. The value can be 
                           "unconstrained", "weak" and "strong".}
\item{lines.Ind}{a set of numbers associated with the steps. The numbers are meaningful only for strong taxonomic hierarchy 
                 where the numbers indicate at which step the depth of taxonomic tree decreases by 1.}
\item{compDat}{The compositional data for performing the PAA.}
\item{taxonomy}{The corresponding taxonomic tree structure of the compositional data if available.}
}
\description{
Conduct hierarchical amalgamation using diversity indexes under the
guidance of taxonomic tree (optional) and produce a set of objects which describe
the hierarchical tree for pairwise amalgamation of components of the compositional 
matrix until there is just a single component.
}
\examples{
data(HIV)
## get the compositional data
compDat <- HIV$compDat[, 1:60]
## taxonomic tree structure
taxonomy <- HIV$taxonomy
## weak taxonomic hierarchy
weak <- hPAA(compDat, taxonomy = taxonomy, method = "Simpson", strong = FALSE)
## strong taxonomic hierarchy
strong <- hPAA(compDat, taxonomy = taxonomy, method = "Simpson", strong = TRUE)
}
