% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHPAA}
\alias{plotHPAA}
\title{Plot Method for the Object with class "hPAA" from hPAA Analysis}
\usage{
plotHPAA(object, taxonomy = TRUE, rank.line = FALSE, 
label = NULL, hang = -1, axes = TRUE, frame.plot = FALSE,
ann = TRUE, lab.cex = 0.8, main = "Dendrogram", sub = NULL, xlab = "",
ylab = "Percentage change (\%)", lim.y = -2, logy = TRUE,
jitters = FALSE, select.level = "all", ...)
}
\arguments{
\item{object}{the output with class "hPAA" from function hPAA().}

\item{taxonomy}{indicate if colored bars representing taxonomic tree structure is added to the dendrogram.}

\item{rank.line}{indicate if horizontal lines showing the taxonomic ranks is added to the dendrogram, 
only valid in the strong hierarchy.}

\item{label}{a character vector of labels for the leaves of the tree.
By default the row names or row numbers of the original data are used.}

\item{hang}{the fraction of the plot height by which labels should hang below the rest of the plot. 
A negative value will cause the labels to hang down from 0.}

\item{axes, frame.plot, ann}{logical flags as in plot.default.}

\item{lab.cex}{size of the lablels for the leaves of the dendrogram}

\item{main, sub, xlab, ylab}{character strings for title.}

\item{lim.y}{lower bound of y for the color plot}

\item{logy}{indicate whether plot the y-axis in log scale.}

\item{jitters}{indicate whether jitters need to be added to make the heights of dendrogram 
more distinguising.}

\item{select.level}{indicate the levels of taxonomic ranks/depths to display on the dendrogram (as horizontal lines).}

\item{...}{further graphical arguments for the colored_bars(). 
E.g., y_shift controls the distance between the labels of dendrogram leaves and color bar.}
}
\description{
Plot the results of hierarchical Principal Amalgamation Analysis (PAA) as 
dendrogram with the y-axis showing the y-axis shows the percentage decrease in total diversity as measured by 
the selected diversity index, optionally include color bars to indicate the taxonomic tree structure of the 
compositional data matrix
}
\examples{
data(HIV)
## get the compositional data
compDat <- HIV$compDat[, 1:60]
## taxonomic tree structure
taxonomy <- HIV$taxonomy
## weak taxonomic hierarchy
weak <- hPAA(compDat, method = "Simpson", taxonomy = taxonomy, strong = FALSE)
plotHPAA(weak, lab.cex = 0.7, ylab = "Percentage change (\%)", 
         y_scale = 1.1, y_shift = -1, cex.rowLabels = 1.1)

}
