% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLine.R
\name{plotLine}
\alias{plotLine}
\title{Plot Method for the Object with class "hPAA" from hPAA Analysis}
\usage{
plotLine(object, 
         xlab = "Number of PCs", 
         ylab = "Percentage change (\%)", 
         main = paste0(object$diversity.method, " Index"), ...)
}
\arguments{
\item{object}{the output with class "hPAA" from function hPAA().}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{main}{a main title for the plot.}

\item{...}{further graphical arguments as in the plot.default.}
}
\description{
Plot the results of hierarchical Principal Amalgamation Analysis (PAA) as 
scree plot, which shows the percentage change in the diversity loss as a 
function of the number of principal compositions.
}
\examples{
data(HIV)
## get the compositional data
compDat <- HIV$compDat[, 1:60]
## taxonomic tree structure
taxonomy <- HIV$taxonomy
## weak taxonomic hierarchy
weak <- hPAA(compDat, method = "Simpson", taxonomy = taxonomy, strong = FALSE)
plotLine(weak)

}
