% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot Method for the Object with class "hPAA" from hPAA Analysis}
\usage{
plotMDS(object, cutoff, ...)
}
\arguments{
\item{object}{the output with class "hPAA" from function hPAA().}

\item{cutoff}{number of principal compostions to cut off from the HPAA path.}

\item{...}{further graphical arguments for the geom_points(). 
E.g., size controls the size of points in the plot.}
}
\description{
Plot the results of hierarchical Principal Amalgamation Analysis (PAA) as 
NMDS ordination plot to visualize the changes in the between-sample distance
patterns before and after HPAA with any given number of principal compositions.
}
\examples{
data(HIV)
## get the compositional data
compDat <- HIV$compDat[, -61]
## taxonomic tree structure
taxonomy <- HIV$taxonomy
## weak taxonomic hierarchy
weak <- hPAA(compDat, method = "Simpson", taxonomy = taxonomy, strong = FALSE)
plotMDS(weak, cutoff = 20)
}
